+++ RayTracing/Ray.h
@@ -6,10 +6,10 @@
 class Ray
 {
 public:
-	glm::vec3 o; // position of light source
-	glm::vec3 d; // direction of light
+	glm::dvec3 o; // position of light source
+	glm::dvec3 d; // direction of light
 	Ray(void);
-	Ray(const glm::vec3& origin, const glm::vec3& dir);
+	Ray(const glm::dvec3& origin, const glm::dvec3& dir);
 	Ray(const Ray& ray);
 
 	Ray& operator= (const Ray& rhs);
